computer_vision_app/pyproject.toml
[project]
name = "computer-vision-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5.3",
    "mediapipe>=0.10.32",
    "opencv-python-headless>=4.11.0.86",
    "python-fasthtml>=0.12.48",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.41.0",
]
